Add files via upload
Raya'an is an E-Application gathers parents, kindergarten/special-needs graduates and daycare in one platform. Parents can see the information of each of the kindergarten / special education graduates, and the daycare. Thus, parents are given multiple choices of Daycares and specialists. Detailed information about those mentioned above are provided as well so parents can choose the best and more comfortable environment that suits them better. As a result, time and efforts will be saved as the communication between these three parties (parents, specialists, and daycares).